requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
pillow
numpy
gdown
